cmake: gcc: target_arm: Force FP hard ABI for DCLS
When Dual-redundant Core Lock-step (DCLS) topology is used, the VFP
registers across the two redundant cores must be manually initialised
and synchronised, and this requires the `-mfloat-abi=hard` option to
be specified.

This commit forces the use of FP "hard" ABI on the VFP-equipped cores
that are configured in DCLS topology.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>
GitOrigin-RevId: 18398948db5444bea0e818dde79865dca990c4b1
Change-Id: Icb6a35de146158de7eda0b38ff0ed500b4fc788f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3629269
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>