Bluetooth: Mesh: Change IV Update duration requirement for new nodes
Mesh Profile specification errata 11627

Section 3.10.5:
When a node is added to a network, the node is given an IV Index.
If the node is added to a network when the network is in Normal
operation, then it shall operate in Normal operation for at least
96 hours. If a node is added to a network while the network is in
the IV Update in Progress state, then the node shall be given the
new IV Index value and operate in IV Update in Progress operation
without the restriction of being in this state for at least 96 hours.

(cherry picked from commit ee3e93148e3a8d0a3f896b9c41b16201c3264f65)

Original-Signed-off-by: Michał Narajowski <michal.narajowski@codecoup.pl>
GitOrigin-RevId: ee3e93148e3a8d0a3f896b9c41b16201c3264f65
Change-Id: Id583ce4896aad0a882f834f06f1aaf1f154527bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4315338
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>