UPSTREAM: dts: npcx7m6fb: define the port as i2c bus rather than the controller
The npcx7m6fb dts defines i2c controller and ports, where the controller
is a property in the port node definition and the device nodes should be
defined as subnodes of the port.

This changes the binding definitions to reflect that, defining the port
as an i2c-controller so that in can inherit sub-nodes for i2c devices,
and the controller as a generic node.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
(cherry picked from commit fa7843d12c363d5c3be95cff6e3040dea4311cbe)
BUG=b:175116395
TEST=build and manually check zephyr.dts and devicetree_unfixed.h
Change-Id: I50641e43f8c043aff6e585043f410d76367ac1cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/2629551
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>