UPSTREAM: drivers: syscon: Add generic syscon API
A syscon device is a device managing a memory region containing a set of
registers that are not cohesive enough to represent as any specific type
of device. We need a driver for that because several other drivers could
use the same region at the same time and we need to io-map the region at
boot for MMU enabled platforms.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>
(cherry picked from commit bc305984565a8f30e8afd02026dc99c16853a362)

BUG=b:179900857
TEST=zmake testall

Change-Id: I7dc7f854502d4e0c581c3b232490fb4095c60aa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3147075
Tested-by: Yuval Peress <peress@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Yuval Peress <peress@chromium.org>