drivers: timer: arm_arch_timer: Workaround for Cortex-A9 erratum 740657
Modification of the ARM architected timer driver and its configuration
data in order to address an erratum which exists at least in the Cor-
tex-A9 CPU, and which can also be observed in the QEMU implementation
of the Cortex-A9.

Comp.: ARM Cortex-A9 processors Software Developer Errata Notice
ARM document ID032315
Erratum 740657

This erratum causes a spurious interrupt pending indication with the
interrupt controller if no new compare value is written within the
timer ISR before the interrupt is cleared. This is usually the case
in tickless mode. If the spurious interrupt is not prevented, the
timer ISR will be called twice, but on second execution, the pending
flag is not set within the timer's register space. Not handling this
issue will lead to erratic tick announcements to the kernel.

(cherry picked from commit 315ad1d6e6d2d62af05a1a213d23a723dcf20771)

Signed-off-by: Immo Birnbaum <Immo.Birnbaum@weidmueller.com>
GitOrigin-RevId: 315ad1d6e6d2d62af05a1a213d23a723dcf20771
Change-Id: I2b1c45d8d4be407d5e7e2cf6876099328ff02b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3253207
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>