drivers: ethernet: xlnx_gem: Zynq-7000 support for the Xilinx GEM driver
Add support for the Xilinx Zynq-7000 SoC family to this driver. This
includes some SoC-specific register accesses when setting an updated
TX clock divider, also, the device tree binding now supports higher
MDC clock divisor values when the current target SoC is a Zynq rather
than a ZynqMP.

With regards to the use of this driver in a QEMU simulation of the
Zynq-7000, the Kconfig file is modified so that the driver is not
enabled unless QEMU networking is set to Ethernet mode.

(cherry picked from commit 266875ead0e81ead2a10179ec8498a683e2e2b32)

Signed-off-by: Immo Birnbaum <Immo.Birnbaum@weidmueller.com>
GitOrigin-RevId: 266875ead0e81ead2a10179ec8498a683e2e2b32
Change-Id: Iab1ffeee5cfa941383b83bfc7c90d11fc638a7c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3253209
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>