soc: intel_asdp: Clean up soc_init() code
Reorganize the initialization code to cleanly separate the platforms
and clarify which code is common.  The #if'ery was sort of a mess.
This is in preparation for an incoming patch that unifies the shim
register definitions across platform variants.

(cherry picked from commit 12df8fca4ed409eb1f26fef626033e9939ccca00)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 12df8fca4ed409eb1f26fef626033e9939ccca00
Change-Id: I5a883a9c5c4407894855a7cce1813d5511104e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300224
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>