soc: intel_adsp: Unify vector table generation
This was an abstraction layer without a purpose.  All existing
platforms have the same (LXn core) layout.  When we need to split this
out in the future, the right thing will be to use the values already
provided by the platform core-isa.h and not duplicate them anyway.

Think of this as a first step to an incoming rework of the Zephyr
Xtensa interrupt entry generation, which is long overdue.

(cherry picked from commit 411d8b96247305429e63647b08213cbc9f0de20b)

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
GitOrigin-RevId: 411d8b96247305429e63647b08213cbc9f0de20b
Change-Id: I1988b8aa5a6a61a0a87b5e9f7b2301d3ff2bb6d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3300229
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>