tests/net: TLS test requires more RX PKT and buffers
Due to the previous change on when to send TCP packet on local IP, pkt
may be held in a queue which is to run on a k_work. This changes the
scheduling, and due to that one of the test is failing to allocate a
RX net_pkt at the time it wants to. (previous TCP connection is not yet
fully closed and still own PKT that new connection cannot get then).

Of course all those waiting paquets require buffers so raising them.

It was verified that there is no leak, adding net_pkt_print() at
tcp_conn_unref() shows that when all tcp connection are finally unrefed:
all net_pkt get freed as well.

(cherry picked from commit 1696f6de35190eb68bd5995caf5f3bf7c947913b)

Signed-off-by: Tomasz Bursztyka <tomasz.bursztyka@linux.intel.com>
GitOrigin-RevId: 1696f6de35190eb68bd5995caf5f3bf7c947913b
Change-Id: Ia75966f9eeb51d5bfba323603c6ff5b16b2a1424
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3262299
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>