drivers: serial: ns16550: fill full fifo
Put the maximum number of bytes into Tx FIFO in the fill_fifo routine
to reduce CPU usage.

Previously, the THRE bit was checked in a loop, but, according to doc -
"In the FIFO mode, it is set when the XMIT FIFO is empty, and is
cleared when at least one byte is written to XMIT FIFO.", so only one
byte was transferred every interrupt. That was generating a big amount
of interrupts, which consumes CPU time.

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
GitOrigin-RevId: 1031cfd51e06b7e1cadbfaeb8de44e952ebbadcf
Change-Id: Ie2e1b6f2cdd8ec22625cb855f7aa4da1b76cfd95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3201190
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>