BACKPORT: drivers: serial: ns16550: fill full fifo
Put the maximum number of bytes into Tx FIFO in the fill_fifo routine
to reduce CPU usage.

Previously, the THRE bit was checked in a loop, but, according to doc -
"In the FIFO mode, it is set when the XMIT FIFO is empty, and is
cleared when at least one byte is written to XMIT FIFO.", so only one
byte was transferred every interrupt. That was generating a big amount
of interrupts, which consumes CPU time.

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
(cherry picked from commit 1031cfd51e06b7e1cadbfaeb8de44e952ebbadcf)

BUG=none
TEST=zmake hayato and it8xxx2_evb

Signed-off-by: Ruibin Chang <Ruibin.Chang@ite.com.tw>
Change-Id: I26f7b55c7da8e1230cdae932ca2481e3680f4be3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3206988
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Dawid Niedźwiecki <dn@semihalf.com>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>