drivers: pinmux stm32F1 remap IP without changing the SWJ_CFG
Change the REMAP bits of the AFIO_MAPR of the stm32F1x soc
with local MACRO without changing the SWJ_CFG (write-only bits).
The serial wire JTAG configuration is taken from the Z_AFIO_REMAP
(value of the CONFIG_GPIO_STM32_SWJ_xxx))
and not read from the MAPR register.
It accesses to the MAPR register directly instead of LL functions.
Note that Remapping on the MAPR2 is not to change.

Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: e763061876c3dfeb140b33bd6a9c1598ffa84e90
Change-Id: I0796586709a127fbe246867d5e924bf10e7cbf44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3206246
Tested-by: Copybara Service <copybara-worker-blackhole@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>