Revert "device: simplify structure of handles array"
This reverts commit ec331c6fe21f2bbbf4b0d1c3667c294320b1c924.

Although it's a valid simplification under the assumption that we're
going to be padding the array out anyway, it would use extra ROM if we
fix the build system issues that are currently forcing gen_handles.py
to introduce extra padding in the handles arrays for linker pass 2.

On the (perhaps optimistic) assumption that we're going to fix the
build system, let's get rid of a commit that would get in the way. The
extra "complexity" in device_required_handles_get() is trivial.

This gets rid of a comment describing the linker passes, but the
structure of the comment is a bit misleading (and it contains
incorrect information for the results of pass 2: the terminator at the
end is DEVICE_HANDLE_ENDS, not DEVICE_HANDLE_NULL).

(cherry picked from commit ea2bbb8b9e81e968726d333b3f3c97d7a4d4bbe5)

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: ea2bbb8b9e81e968726d333b3f3c97d7a4d4bbe5
Change-Id: I8a7451138eb665e4a6e24401b7fcccaa35236571
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3227738
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>