interrupt_controller: gicv3_its: Implement Indirect Two-level table
When the ITS is configured with a large number of DeviceID bits,
up to 32, the Device Translation Table can be configured as
"indirect" mode consisting of a 2-level table with the first
level entries pointing to a second level table for each
group of DeviceIDs.

This is necessary to support platforms with DeviceID bits > 16.

(cherry picked from commit 71766f5c050e8821dad7cb452460e83baf3971e1)

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
GitOrigin-RevId: 71766f5c050e8821dad7cb452460e83baf3971e1
Change-Id: If2ae32280a83571bbb57680effa196f995ebbaba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3236983
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>