tracing: Automatic syscall tracing
When generating syscall wrappers, call a tracing macro with the id,
name, and all parameters of the syscall as params when entering and
leaving the syscall. This can be disabled in certain call sites
by defining DISABLE_SYSCALL_TRACING which is useful for certain
tracing implementations which require syscalls themselves to work.

Notably some syscalls *cannot* be automatically traced this way and
headers where exclusions are set are in the gen_syscall.py as notracing.

Includes a systemview and test format implementation.

Tested with systemview, usb, and uart backends with the string
formatter using the tracing sample app.

Debugging the trace wrapper can be aided by setting the TRACE_DIAGNOSTIC
env var and rebuilding from scratch, a warning is issued for every
instance a syscall is traced.

Automatically generating a name mapping for SYSVIEW_Zephyr.txt is a
future item as is documenting how to capture and use the tracing data
generated.

(cherry picked from commit 97dc88bb6d0b3624550edd5142c84a1f36696d24)

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: 97dc88bb6d0b3624550edd5142c84a1f36696d24
Change-Id: I35b8845d5cb37c3894a97348352d6e83e5da65cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3238394
Tested-by: CopyBot <cros.copybot.ota@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>