drivers: mcux_gpt_timer: Enabled GPT timer on MIMXRT1064 EVK
This enables the GPT timer for use as a hardware clock on the
MIMXRT1064. The timer will be disabled when builds use the GPT timer as
a counter, and the build will fall back to using the Cortex M systick as
a timer. This was tested using the kernel tickless test, to verify the
driver functions as expected.

(cherry picked from commit e99eb6bb70f0a9fd55092b2eef5702f16ce7a6c5)

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: e99eb6bb70f0a9fd55092b2eef5702f16ce7a6c5
Change-Id: If61405f4d11168816dc4314374edf31c0829b920
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3338817
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>